backend/pyproject.toml
[project]
name = "gambling-search-api"
version = "0.1.0"
description = "불법 도박 사이트 자동 검색/채증 시스템 - FastAPI 백엔드"
requires-python = ">=3.11"
dependencies = [
    # FastAPI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.9",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    # Redis
    "redis>=5.2.0",
    # Storage (MinIO / S3)
    "boto3>=1.35.0",
    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",
    # UUID v7
    "uuid6>=2024.7.10",
    # HTTP Client
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
